seperated connect contact card from lead details page
